APR-2402 moved switch to new controller
